Implement language updates and social media links in user profiles

+++ database/migrations/2025_07_09_120803_add_is_public_to_projects_table.php
@@ -8,14 +8,20 @@ return new class extends Migration {
     public function up()
     {
         Schema::table('projects', function (Blueprint $table) {
-            $table->boolean('is_public')->default(false)->after('id');
+            //if not exists, add is_public column
+            if (!Schema::hasColumn('projects', 'is_public')) {
+                $table->boolean('is_public')->default(false)->after('id');
+            }
         });
     }
 
     public function down()
     {
         Schema::table('projects', function (Blueprint $table) {
-            $table->dropColumn('is_public');
+            //if exists, drop is_public column
+            if (Schema::hasColumn('projects', 'is_public')) {
+                $table->dropColumn('is_public');
+            }
         });
     }
 };
